requirements.txt
pandas
scikit-learn
joblib
discord.py
Sastrawi
python-dotenv